Adding unit tests and necessary references
A reference to the `Moq` package has been added to the `TestApiTodoApp.csproj` file to allow the creation of mock objects in unit tests.

A reference to the `TodoApp.API` project has been added to the `TestApiTodoApp.csproj` file to allow access to its classes and methods from the test project.